data-science/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
psycopg2-binary>=2.9.0
